pyproject.toml
[tool.poetry]
name = "crud_flask"
version = "0.1.0"
description = ""
authors = ["xandowski <xandy-moraes@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
python-dotenv = "^0.19.2"
Flask-Cors = "^3.0.10"
six = "^1.16.0"
wheel = "^0.37.1"
Authlib = "^0.15.5"
Flask-SQLAlchemy = "^2.5.1"
mysql-connector-python = "^8.0.28"
Flask-Migrate = "^3.1.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
